ai_business_assistant/requirements.txt
# Core runtime + test dependencies for Business AI Assistant.
#
# Heavy/optional ML dependencies (transformers/torch/tensorflow/prophet/etc.) have been
# intentionally excluded from this base set to keep installs fast and reliable.
# Modules that can leverage those libraries should import them lazily and degrade
# gracefully when they're not available.

# API / microservices
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
slowapi==0.1.9
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Data / persistence
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.13.1
redis==5.2.1
celery==5.4.0

# Security
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
python-multipart>=0.0.6,<0.0.7

# Observability
prometheus-client==0.21.1
structlog==23.2.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-sqlalchemy==0.45b0
opentelemetry-instrumentation-redis==0.45b0
opentelemetry-instrumentation-celery==0.45b0
opentelemetry-exporter-otlp==1.24.0

# Analytics used by unit tests
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
textblob>=0.17.1

# Warehouse / ETL
duckdb>=1.0.0
requests
tenacity

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
